+++ src/screens/Home/Messages.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { View, Text, Image } from "react-native";
+import { View, Text, Image, TouchableOpacity } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import MessageBox from "../../components/MessageBox";
 import SearchBox from "../../components/SearchBox";
@@ -10,7 +10,7 @@ import messagesStyles from "./utils/messagesStyles";
 
 const profileImg = require("../../assets/profile-picture.png");
 
-const Messages = () => {
+const Messages = ({ navigation }) => {
   const [searchQuery, setSearchQuery] = useState("");
   const indicator = useIndicator("openToChat");
   const bgImg = useBackground("topBubbles");
@@ -30,7 +30,10 @@ const Messages = () => {
             <Text style={messagesStyles.customMessage}>Set Custom Message</Text>
           </View>
         </View>
-        <Image style={messagesStyles.pencilIcon} source={pencilIcon} />
+
+        <TouchableOpacity onPress={() => navigation.navigate("NewMessage")}>
+          <Image style={messagesStyles.pencilIcon} source={pencilIcon} />
+        </TouchableOpacity>
       </View>
       <View style={messagesStyles.messageContainer}>
         <Text style={messagesStyles.tabLabel}>Messages</Text>