+++ CHANGELOG.md
@@ -9,6 +9,7 @@ NOTE: all lowercase and no ending punctuation marks
 
 ## Next release v1.0
 
+- [LE] - [#16](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/16) app(screen): new message screen
 - [CY] - [#11](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/11) app(screen): messages screen
 - [CY] - [#9](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/9) app(feature): home bottom tabs navigation
 - [CY] - [#7](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/7) app(feature): add dependencies to enable expo to publish the app online